backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic[email]==2.10.3
openai==1.57.2
chromadb==0.5.23
pdfplumber==0.11.4
python-docx==1.1.2
httpx==0.28.1
aiofiles==24.1.0
bcrypt==4.2.1
PyJWT==2.10.1
email-validator==2.2.0
